Refactor tests slightly
